added voxelization script
